+++ src/util/point.rs
@@ -15,19 +15,19 @@ pub const DIAGONAL: [Point; 4] = [
 ];
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
 pub struct Point {
-    pub x: i32,
-    pub y: i32,
+    pub x: i64,
+    pub y: i64,
 }
 
 impl Point {
     #[inline]
-    pub const fn new(x: i32, y: i32) -> Self {
+    pub const fn new(x: i64, y: i64) -> Self {
         Self { x, y }
     }
     pub const fn from_usize(x: usize, y: usize) -> Self {
         Self {
-            x: x as i32,
-            y: y as i32,
+            x: x as i64,
+            y: y as i64,
         }
     }
     #[inline]
@@ -39,7 +39,7 @@ impl Point {
         Point::new(self.y, -self.x)
     }
     #[inline]
-    pub fn manhattan(self, other: Self) -> u32 {
+    pub fn manhattan(self, other: Self) -> u64 {
         self.x.abs_diff(other.x) + self.y.abs_diff(other.y)
     }
 }
@@ -57,17 +57,17 @@ impl From<u8> for Point {
     }
 }
 
-impl Mul<i32> for Point {
+impl Mul<i64> for Point {
     type Output = Self;
     #[inline]
-    fn mul(self, rhs: i32) -> Self::Output {
+    fn mul(self, rhs: i64) -> Self::Output {
         Point::new(self.x * rhs, self.y * rhs)
     }
 }
 
-impl MulAssign<i32> for Point {
+impl MulAssign<i64> for Point {
     #[inline]
-    fn mul_assign(&mut self, rhs: i32) {
+    fn mul_assign(&mut self, rhs: i64) {
         self.x *= rhs;
         self.y *= rhs;
     }
